update SgfcMessage to use unsigned long for line/column numbers (fixes #40)

+++ include/ISgfcMessage.h
@@ -76,13 +76,13 @@ namespace LibSgfcPlusPlus
     /// that caused the message. Returns SgfcConstants::InvalidLineNumber if
     /// the message does not refer to a specific line in the SGF data (most
     /// fatal errors, but also some errors and warnings).
-    virtual int GetLineNumber() const = 0;
+    virtual unsigned long GetLineNumber() const = 0;
 
     /// @brief Returns the number of the column (1-based) in the parsed SGF data
     /// that caused the message. Returns SgfcConstants::InvalidColumnNumber if
     /// the message does not refer to a specific column in the SGF data (most
     /// fatal errors, but also some errors and warnings).
-    virtual int GetColumnNumber() const = 0;
+    virtual unsigned long GetColumnNumber() const = 0;
 
     /// @brief Returns true if the message describes a critical problem. Only
     /// warning and error messages can be critical. Always returns false for
